% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-CircularBlocks.R
\docType{class}
\name{CircularBlocks-class}
\alias{CircularBlocks}
\alias{CircularBlocks-class}
\title{Class for Circular Blocks Bootstrap implementation.}
\description{
\code{CircularBlocks} is an S4 class that implements the moving blocks
bootstrap described in Politis and Romano (1992).
}
\details{
\code{CircularBlocks} extends the S4 class
\code{\link{BootPos}} and the remarks made in its documentation
apply here as well.

The Circular Blocks Bootstrap method of Politis and Romano (1992) resamples blocks
randomly, with replacement from the collection of overlapping blocks of
length \code{l} that start with observation 1, 2, \ldots, \code{N}.
In contrast to \code{\link{MovingBlocks}} and \code{\link{NonoverlappingBlocks}},
the Circular Blocks Bootstrap uses elements from the periodically extended series.
A more precise description of the procedure can also be found in
Lahiri (1999), p. 389.
}
\references{
Politis, D. and Romano, J. P. (1992). A circular block resampling procedure
for stationary data. In \emph{Exploring the Limits of Bootstrap}
(R. Lepage and L. Billar, eds.), 263--270. Wiley, New York.
}
\seealso{
\code{\link{getPositions-CircularBlocks}}
}
\keyword{S4-classes}

