% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-CircularBlocks.R
\name{CircularBlocks-constructor}
\alias{CircularBlocks-constructor}
\alias{circularBlocks}
\title{Create an instance of the \code{\link{CircularBlocks}} class.}
\usage{
circularBlocks(l, N)
}
\arguments{
\item{l}{the block length for the block bootstrap methods}

\item{N}{number of available observations to bootstrap from}
}
\value{
Returns an instance of \code{CircularBlocks}.
}
\description{
Create an instance of the \code{\link{CircularBlocks}} class.
}
\keyword{Constructors}

