% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-CovarianceMatrixMultipliers.R
\docType{class}
\encoding{latin1}
\name{CovarianceMatrixMultipliers-class}
\alias{CovarianceMatrixMultipliers}
\alias{CovarianceMatrixMultipliers-class}
\title{Class for Covariance Matrix Multiplier implementation.}
\description{
\code{CovarianceMatrixMultipliers} is an S4 class that implements the moving blocks
bootstrap described in B{\"u}cher and Kojadinovic (2016), Section 5.2.2.
}
\details{
\code{CovarianceMatrixMultipliers} extends the S4 class
\code{\link{DependentMultipliers}} and the remarks made in its documentation
apply here as well.

TODO: Add more description here!
}
\section{Slots}{

\describe{
\item{\code{phi}}{function from assumption (M3)}
}}
\references{
B{\"u}cher, A., & Kojadinovic, I. (2016).
A dependent multiplier bootstrap for the sequential empirical copula
process under strong mixing. \emph{Bernoulli}, \bold{22}(2), 927--968.
}
\seealso{
\code{\link{getMultipliers-CovarianceMatrixMultipliers}}
}
\keyword{S4-classes}

