% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generics-associations}
\alias{generics-associations}
\alias{getQuantilePG}
\alias{getBootPos}
\alias{getFreqRep}
\alias{getQuantileSD}
\alias{getWeight}
\title{Generic functions for accessing associations of objects}
\usage{
getQuantilePG(object, ...)

getBootPos(object, ...)

getFreqRep(object, ...)

getQuantileSD(object, ...)

getWeight(object, ...)
}
\arguments{
\item{object}{object from which to get the associated object}

\item{...}{optional parameters; for documentation see the documentation of
the methods to each of the generic.}
}
\description{
These generic functions are needed to access the objects' associated objects.
Note that the naming convention \code{getAssociatedObject} was applied, where
\code{AssociatedObject} is the name of the class of the associated object.
}
\seealso{
For an overview on the classes of the framework, and all
associations, see the class diagrams in the package description
[cf. \code{\link{quantspec-package}}].
}
