\name{generics-functions}
\alias{generics-functions}
\alias{getPositions}
\alias{increasePrecision}
\title{Generic functions for implementation of methods of a class}
\usage{
increasePrecision(object, ...)

getPositions(object, ...)
}
\arguments{
  \item{object}{specifies the object from which the method
  is to be applied.}

  \item{...}{optional parameters; for documentation see the
  documentation of the methods to the generic.}
}
\description{
These generic functions need to be defined to allow for the
automatic dispaching mechanism.
}
\seealso{
For an overview on the classes of the framework, and all of
their methods, see the class diagrams in the package
description [cf. \code{\link{quantspec-package}}].
}

