% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-FreqRep.R
\name{getB-FreqRep}
\alias{getB-FreqRep}
\alias{getB,FreqRep-method}
\title{Get \code{B} from a \code{\link{FreqRep}} object.}
\usage{
\S4method{getB}{FreqRep}(object)
}
\arguments{
\item{object}{\code{FreqRep} of which to get the \code{B}}
}
\value{
Returns the attribute \code{B} that's a slot of \code{object}.
}
\description{
Get \code{B} from a \code{\link{FreqRep}} object.
}
\keyword{Access-functions}
