% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-LagOperator.R
\name{getB-LagOperator}
\alias{getB-LagOperator}
\alias{getB,LagOperator-method}
\title{Get \code{B} from a \code{\link{LagOperator}} object.}
\usage{
\S4method{getB}{LagOperator}(object)
}
\arguments{
\item{object}{\code{LagOperator} of which to get the \code{B}}
}
\value{
Returns the attribute \code{B} that's a slot of \code{object}.
}
\description{
Get \code{B} from a \code{\link{LagOperator}} object.
}
\keyword{Access-functions}
