% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-FreqRep.R
\docType{methods}
\name{getBootMultipliers-FreqRep}
\alias{getBootMultipliers,FreqRep-method}
\alias{getBootMultipliers-FreqRep}
\title{Get associated \code{\link{BootMultipliers}} from a
\code{\link{FreqRep}}.}
\usage{
\S4method{getBootMultipliers}{FreqRep}(object)
}
\arguments{
\item{object}{\code{FreqRep} from which to get the
\code{\link{BootMultipliers}}.}
}
\value{
Returns the \code{\link{BootMultipliers}} object associated.
}
\description{
Get associated \code{\link{BootMultipliers}} from a
\code{\link{FreqRep}}.
}
\keyword{Access-association-functions}

