% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-LagOperator.R
\name{getBootPos-LagOperator}
\alias{getBootPos-LagOperator}
\alias{getBootPos,LagOperator-method}
\title{Get associated \code{\link{BootPos}} from a
\code{\link{LagOperator}}.}
\usage{
\S4method{getBootPos}{LagOperator}(object)
}
\arguments{
\item{object}{\code{LagOperator} from which to get the
\code{\link{BootPos}}.}
}
\value{
Returns the \code{\link{BootPos}} object associated.
}
\description{
Get associated \code{\link{BootPos}} from a
\code{\link{LagOperator}}.
}
\keyword{Access-association-functions}
