% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-SmoothedPG.R
\docType{methods}
\name{getCoherencePointwiseCIs-SmoothedPG}
\alias{getCoherencePointwiseCIs,SmoothedPG-method}
\alias{getCoherencePointwiseCIs-SmoothedPG}
\title{Get pointwise confidence intervals for the quantile coherence}
\usage{
\S4method{getCoherencePointwiseCIs}{SmoothedPG}(object, frequencies = 2 * pi *
  (0:(lenTS(object@qPG@freqRep@Y) - 1))/lenTS(object@qPG@freqRep@Y),
  levels.1 = getLevels(object, 1), levels.2 = getLevels(object, 2),
  d1 = 1:(dim(object@values)[2]), d2 = 1:(dim(object@values)[4]),
  alpha = 0.1, type = c("naive.sd", "variance stabilized", "boot.sd",
  "boot.full"))
}
\arguments{
\item{object}{\code{SmoothedPG} of which to get the confidence intervals}

\item{frequencies}{a vector of frequencies for which to get the result}

\item{levels.1}{the first vector of levels for which to get the result}

\item{levels.2}{the second vector of levels for which to get the result}

\item{d1}{optional parameter that determine for which j1 to return the
data; may be a vector of elements 1, ..., D}

\item{d2}{same as d1, but for j2}

\item{alpha}{the level of the confidence interval; must be from \eqn{(0,1)}}

\item{type}{a flag indicating which type of confidence interval should be
        returned; can take one of the three values discussed above.}
}
\value{
Returns a named list of two arrays \code{lowerCIS} and \code{upperCIs}
         containing the lower and upper bounds for the confidence intervals.
}
\description{
TODO
}
\keyword{Access-functions}

