\docType{methods}
\name{getDescr-Weight}
\alias{getDescr,Weight-method}
\alias{getDescr-Weight}
\title{Get attribute \code{descr} from a \code{Weight}.}
\usage{
\S4method{getDescr}{Weight}(object)
}
\arguments{
  \item{object}{\code{Weight} from which to get the
  \code{descr}.}
}
\value{
Returns the \code{descr} attribute.
}
\description{
Get attribute \code{descr} from a \code{Weight}.
}
\keyword{Access-functions}

