% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/Class-FreqRep.R
\docType{methods}
\name{getFrequencies-FreqRep}
\alias{getFrequencies,FreqRep-method}
\alias{getFrequencies-FreqRep}
\title{Get attribute \code{frequencies} from a \code{\link{FreqRep}}.}
\usage{
\S4method{getFrequencies}{FreqRep}(object)
}
\arguments{
\item{object}{\code{FreqRep} from which to get the
        \code{frequencies}.}
}
\value{
Returns the \code{frequencies} attribute, as a vector of real numbers.
}
\description{
Get attribute \code{frequencies} from a \code{\link{FreqRep}}.
}
\keyword{Access-functions}

