% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-LagOperator.R
\name{getIsRankBased-LagOperator}
\alias{getIsRankBased-LagOperator}
\alias{getIsRankBased,LagOperator-method}
\title{Get \code{isRankBased} from a \code{\link{LagOperator}} object}
\usage{
\S4method{getIsRankBased}{LagOperator}(object)
}
\arguments{
\item{object}{\code{LagOperator} of which to get the \code{isRankBased}}
}
\value{
Returns the attribute \code{isRankBased} that's a slot of \code{object}.
}
\description{
Get \code{isRankBased} from a \code{\link{LagOperator}} object
}
\keyword{Access-functions}
