% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-LagEstimator.R
\name{getLagOperator-LagEstimator}
\alias{getLagOperator-LagEstimator}
\alias{getLagOperator,LagEstimator-method}
\title{Get associated \code{\link{LagOperator}} from a \code{\link{LagEstimator}}.}
\usage{
\S4method{getLagOperator}{LagEstimator}(object)
}
\arguments{
\item{object}{\code{LagEstimator} from which to get the \code{LagOperator}.}
}
\value{
Returns the \code{\link{LagOperator}} object associated.
}
\description{
Get associated \code{\link{LagOperator}} from a \code{\link{LagEstimator}}.
}
\keyword{Access-association-functions}
