% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-LagOperator.R
\docType{methods}
\name{getLevels-LagOperator}
\alias{getLevels,LagOperator-method}
\alias{getLevels-LagOperator}
\title{Get attribute \code{levels} from a \code{LagOperator}.}
\usage{
\S4method{getLevels}{LagOperator}(object, j)
}
\arguments{
\item{object}{\code{LagOperator} from which to get the \code{levels}.}

\item{j}{Index pointing to a set of levels in the list; optional.}
}
\value{
Returns levels attribute, as a vector of real numbers.
}
\description{
If the optional parameter \code{j} is supplied, then the \code{j}th vector of
levels will be returned, a list with all vectors otherwise.
}
\keyword{Access-functions}

