% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-QSpecQuantity.R
\docType{methods}
\name{getLevels-QSpecQuantity}
\alias{getLevels,QSpecQuantity-method}
\alias{getLevels-QSpecQuantity}
\title{Get attribute \code{levels} from a \code{QSpecQuantity}.}
\usage{
\S4method{getLevels}{QSpecQuantity}(object, j)
}
\arguments{
\item{object}{\code{QSpecQuantity} from which to get the \code{levels}.}

\item{j}{Index pointing to a set of levels in the list; optional.}
}
\value{
Returns levels attribute, as a vector of real numbers.
}
\description{
If the optional parameter \code{j} is supplied, then the \code{j}th vector of
levels will be returned, a list with all vectors otherwise.
}
\examples{
qPG         <- quantilePG(rnorm(10), levels.1=c(0.25,0.5))
levels.list <- getLevels(qPG)
levels.1    <- getLevels(qPG,1)
}
\keyword{Access-functions}

