% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-CovarianceMatrixMultipliers.R
\docType{methods}
\name{getMultipliers-CovarianceMatrixMultipliers}
\alias{getMultipliers,CovarianceMatrixMultipliers-method}
\alias{getMultipliers-CovarianceMatrixMultipliers}
\title{Get Multipliers for the Covariance Matrix Multiplier.}
\usage{
\S4method{getMultipliers}{CovarianceMatrixMultipliers}(object, B = 1)
}
\arguments{
\item{object}{a \code{CovarianceMatrixMultipliers} object; used to specify the
parameters \code{N}, \code{l}, \code{phi} and the type of the
scheme to generate the dependent multipliers.}

\item{B}{Number of independent repetitions to bootstrap.}
}
\value{
a matrix of dimension \code{[N,B]} where each column gives a sequence
        of multipliers generated according to the moving average scheme.
}
\description{
Get Multipliers for the Covariance Matrix Multiplier.
}

