% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-MovingAverageMultipliers.R
\docType{methods}
\name{getMultipliers-MovingAverageMultipliers}
\alias{getMultipliers,MovingAverageMultipliers-method}
\alias{getMultipliers-MovingAverageMultipliers}
\title{Get Multipliers for the Moving Average Multiplier.}
\usage{
\S4method{getMultipliers}{MovingAverageMultipliers}(object, B = 1)
}
\arguments{
\item{object}{a \code{MovingAverageMultipliers} object; used to specify the
parameters \code{N}, \code{l}, \code{phi} and the type of the
scheme to generate the dependent multipliers.}

\item{B}{Number of independent repetitions to bootstrap.}
}
\value{
a matrix of dimension \code{[N,B]} where each column gives a sequence
        of multipliers generated according to the moving average scheme.
}
\description{
Get Multipliers for the Moving Average Multiplier.
}

