% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-NoneMultipliers.R
\docType{methods}
\name{getMultipliers-NoneMultipliers}
\alias{getMultipliers,NoneMultipliers-method}
\alias{getMultipliers-NoneMultipliers}
\title{Get Multipliers for the None Multiplier.}
\usage{
\S4method{getMultipliers}{NoneMultipliers}(object, B = 1)
}
\arguments{
\item{object}{a \code{NoneMultipliers} object; used to specify the
parameters \code{N}, \code{l} and the type of the
scheme to generate the dependent multipliers.}

\item{B}{Number of independent repetitions to bootstrap.}
}
\value{
a matrix of dimension \code{[N,B]} with all 1.
}
\description{
Get Multipliers for the None Multiplier.
}

