% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-QuantileSD.R
\docType{methods}
\name{getN-QuantileSD}
\alias{getN,QuantileSD-method}
\alias{getN-QuantileSD}
\title{Get \code{N} from a quantile spectral density kernel}
\usage{
\S4method{getN}{QuantileSD}(object)
}
\arguments{
\item{object}{\code{QuantileSD} of which to get the \code{N}}
}
\value{
Returns the attribute \code{N} that's a slot of \code{object}.
}
\description{
Get \code{N} from a quantile spectral density kernel
}
\keyword{Access-functions}

