% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-QRegEstimator.R
\name{getParallel-QRegEstimator}
\alias{getParallel-QRegEstimator}
\alias{getParallel,QRegEstimator-method}
\title{Get \code{getParallel} from a \code{\link{QRegEstimator}} object}
\usage{
\S4method{getParallel}{QRegEstimator}(object)
}
\arguments{
\item{object}{\code{QRegEstimator} of which to get the \code{parallel}}
}
\value{
Returns the attribute \code{parallel} that's a slot of \code{object}.
}
\description{
Get \code{getParallel} from a \code{\link{QRegEstimator}} object
}
\keyword{Access-functions}
