% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-StationaryBlocks.R
\docType{methods}
\name{getPositions-StationaryBlocks}
\alias{getPositions,StationaryBlocks-method}
\alias{getPositions-StationaryBlocks}
\title{Get Positions for the Stationary Bootstrap.}
\usage{
\S4method{getPositions}{StationaryBlocks}(object, B = 1)
}
\arguments{
\item{object}{a \code{StationaryBlocks} object; used to specify the parameters
\code{N}, \code{l} and the type of the bootstrap.}

\item{B}{Number of independent repetitions to bootstrap.}
}
\value{
a matrix of dimension \code{[N,B]} where each column gives the
        positions in which to reorder the observations to yield one
         bootstrap replication.
}
\description{
Get Positions for the Stationary Bootstrap.
}

