\docType{methods}
\name{getQuantilePG-QuantileSD}
\alias{getQuantilePG,QuantileSD-method}
\alias{getQuantilePG-QuantileSD}
\title{Get associated \code{\link{QuantilePG}} from a \code{\link{QuantileSD}}.}
\usage{
\S4method{getQuantilePG}{QuantileSD}(object)
}
\arguments{
  \item{object}{\code{QuantileSD} from which to get the
  \code{\link{QuantilePG}}.}
}
\value{
Returns the \code{\link{QuantilePG}} object associated.
}
\description{
Get associated \code{\link{QuantilePG}} from a
\code{\link{QuantileSD}}.
}
\keyword{Access-association-functions}

