% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-IntegrQuantileSD.R
\docType{methods}
\name{getQuantileSD-IntegrQuantileSD}
\alias{getQuantileSD,IntegrQuantileSD-method}
\alias{getQuantileSD-IntegrQuantileSD}
\title{Get associated \code{\link{getQuantileSD}} from an
\code{\link{IntegrQuantileSD}}.}
\usage{
\S4method{getQuantileSD}{IntegrQuantileSD}(object)
}
\arguments{
\item{object}{\code{IntegrQuantileSD} from which to get the
\code{\link{getQuantileSD}}.}
}
\value{
Returns the \code{\link{getQuantileSD}} object associated.
}
\description{
Get associated \code{\link{getQuantileSD}} from an
\code{\link{IntegrQuantileSD}}.
}
\keyword{Access-association-functions}

