% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-QuantileSD.R
\docType{methods}
\name{getR-QuantileSD}
\alias{getR,QuantileSD-method}
\alias{getR-QuantileSD}
\title{Get \code{R} from a quantile spectral density kernel}
\usage{
\S4method{getR}{QuantileSD}(object)
}
\arguments{
\item{object}{\code{QuantileSD} of which to get the \code{R}}
}
\value{
Returns the attribute \code{R} that's a slot of \code{object}.
}
\description{
Get \code{R} from a quantile spectral density kernel
}
\keyword{Access-functions}

