% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-QuantileSD.R
\name{getTs-QuantileSD}
\alias{getTs-QuantileSD}
\alias{getTs,QuantileSD-method}
\title{Get \code{ts} from a quantile spectral density kernel}
\usage{
\S4method{getTs}{QuantileSD}(object)
}
\arguments{
\item{object}{\code{QuantileSD} of which to get the \code{ts}}
}
\value{
Returns the attribute \code{ts} that's a slot of \code{object}.
}
\description{
Get \code{ts} from a quantile spectral density kernel
}
\keyword{Access-functions}
