\docType{methods}
\name{getType-QuantileSD}
\alias{getType,QuantileSD-method}
\alias{getType-QuantileSD}
\title{Get \code{type} from a quantile spectral density kernel}
\usage{
\S4method{getType}{QuantileSD}(object)
}
\arguments{
  \item{object}{\code{QuantileSD} of which to get the
  \code{type}}
}
\value{
Returns the attribute \code{type} that's a slot of
\code{object}.
}
\description{
Get \code{type} from a quantile spectral density kernel
}
\keyword{Access-functions}

