% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-LagOperator.R
\docType{methods}
\name{getValues-LagOperator}
\alias{getValues,LagOperator-method}
\alias{getValues-LagOperator}
\title{Get attribute \code{values} from a \code{LagOperator}.}
\usage{
\S4method{getValues}{LagOperator}(object, levels.1, levels.2)
}
\arguments{
\item{object}{\code{LagOperator} from which to get the \code{values}.}

\item{levels.1}{the first vector of levels for which to get the values}

\item{levels.2}{the second vector of levels for which to get the values}
}
\value{
Returns the \code{values} attribute.
}
\description{
Get attribute \code{values} from a \code{LagOperator}.
}
\keyword{Access-functions}

