% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-KernelWeight.R
\name{getW-KernelWeight}
\alias{getW-KernelWeight}
\alias{getW,KernelWeight-method}
\title{Get attribute \code{W} (kernel used for smoothing) from a \code{KernelWeight}.}
\usage{
\S4method{getW}{KernelWeight}(object)
}
\arguments{
\item{object}{\code{KernelWeight} from which to get the kernel \code{W}.}
}
\value{
Returns the \code{W} attribute.
}
\description{
Get attribute \code{W} (kernel used for smoothing) from a \code{KernelWeight}.
}
\keyword{Access-functions}
