% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-LagKernelWeight.R
\docType{methods}
\name{getW-LagKernelWeight}
\alias{getW,LagKernelWeight-method}
\alias{getW-LagKernelWeight}
\title{Get attribute \code{W} (kernel used for smoothing) from a \code{LagKernelWeight}.}
\usage{
\S4method{getW}{LagKernelWeight}(object)
}
\arguments{
\item{object}{\code{LagKernelWeight} from which to get the kernel \code{W}.}
}
\value{
Returns the \code{W} attribute.
}
\description{
Get attribute \code{W} (kernel used for smoothing) from a \code{LagKernelWeight}.
}
\keyword{Access-functions}

