% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-LagEstimator.R
\name{getWeight-LagEstimator}
\alias{getWeight-LagEstimator}
\alias{getWeight,LagEstimator-method}
\title{Get associated \code{\link{Weight}} from a \code{\link{LagEstimator}}.}
\usage{
\S4method{getWeight}{LagEstimator}(object)
}
\arguments{
\item{object}{\code{LagEstimator} from which to get the \code{Weight}.}
}
\value{
Returns the \code{\link{Weight}} object associated.
}
\description{
Get associated \code{\link{Weight}} from a \code{\link{LagEstimator}}.
}
\keyword{Access-association-functions}
