% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SmoothedPG.R
\docType{methods}
\name{getWeight-SmoothedPG}
\alias{getWeight,SmoothedPG-method}
\alias{getWeight-SmoothedPG}
\title{Get associated \code{\link{Weight}} from a \code{\link{SmoothedPG}}.}
\usage{
\S4method{getWeight}{SmoothedPG}(object)
}
\arguments{
\item{object}{\code{SmoothedPG} from which to get the \code{Weight}.}
}
\value{
Returns the \code{\link{Weight}} object associated.
}
\description{
Get associated \code{\link{Weight}} from a \code{\link{SmoothedPG}}.
}
\keyword{Access-association-functions}

