\docType{methods}
\name{getY-FreqRep}
\alias{getY,FreqRep-method}
\alias{getY-FreqRep}
\title{Get \code{Y} from a \code{\link{FreqRep}} object.}
\usage{
\S4method{getY}{FreqRep}(object)
}
\arguments{
  \item{object}{\code{FreqRep} of which to get the
  \code{Y}}
}
\value{
Returns the attribute \code{Y} that's a slot of
\code{object}.
}
\description{
Get \code{Y} from a \code{\link{FreqRep}} object.
}

