% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernels-multiplier}
\alias{kappaB}
\alias{kappaF}
\alias{kappaP}
\alias{kappaT}
\alias{kernels-multiplier}
\title{Further Kernel function, used for the dependent multiplier sequences}
\usage{
kappaT(x)

kappaB(x)

kappaP(x)

kappaF(x, c = 0.14)
}
\arguments{
\item{x}{real-valued argument to the function; can be a vector}

\item{c}{cut-off point, \eqn{0 \leq c \leq 1}.}
}
\description{
Implementations of kernel functions
}
\details{
Truncated \code{kappaT}:
\deqn{\kappa_T(x) = I\{|x| \leq 1\}.}

Bartlett kernel \code{kappaB}:
\deqn{(1 - |x|) \vee 0.}

Parzen kernel \code{kappaP}:
\deqn{(1-6x^2+6|x|^3)I\{|x| \leq 1/2\} + 2 (1-|x|)^3 I\{1/2 < |x| \leq 1\}.}

Flat top kernel \code{kappaF}:
\deqn{[\{(1 - |x|)/(1-c)\} \vee 0] \wedge 1, \quad 0 \leq c \leq 1.}
}
\examples{
plot(x=seq(-2,2,0.05), y=kappaT(seq(-2,2,0.05)), type="l")
plot(x=seq(-2,2,0.05), y=kappaB(seq(-2,2,0.05)), type="l")
plot(x=seq(-2,2,0.05), y=kappaP(seq(-2,2,0.05)), type="l")
plot(x=seq(-2,2,0.05), y=kappaF(seq(-2,2,0.05)), type="l")
plot(x=seq(-2,2,0.05), y=kappaF(seq(-2,2,0.05), 0.4), type="l")
}

