% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-functions.R
\name{lenTS}
\alias{lenTS}
\title{Validates if \code{Y} is of an appropriate type for a time series and
returns the length of the time series.}
\usage{
lenTS(Y)
}
\arguments{
\item{Y}{the time series to be validated and of which the length is to
be returned.}
}
\value{
Returns the length of the time series after validating it's valid.
}
\description{
Runs \code{\link{timeSeriesValidator}} and returns the number of rows of the
returned matrix.
}
\examples{
Y <- lenTS(sp500)
Y <- lenTS(wheatprices)
Y <- lenTS(rnorm(10))
\dontrun{Y <- lenTS("Not a valid input")}
}
\keyword{Validator-functions}

