% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-MovingAverageMultipliers.R
\docType{class}
\encoding{latin1}
\name{MovingAverageMultipliers-class}
\alias{MovingAverageMultipliers}
\alias{MovingAverageMultipliers-class}
\title{Class for Moving Average Multiplier implementation.}
\description{
\code{MovingAverageMultipliers} is an S4 class that implements the moving blocks
bootstrap described in B{\"u}cher and Kojadinovic (2016), Section 5.2.1.
}
\details{
\code{MovingAverageMultipliers} extends the S4 class
\code{\link{DependentMultipliers}} and the remarks made in its documentation
apply here as well.

TODO: Add more description here!
}
\section{Slots}{

\describe{
\item{\code{kappa}}{function from Section 5.2.1 (B{\"u}cher and Kojadinovic, 2016)}

\item{\code{distrInnov}}{function with one argument n that generates the independent
innovations}
}}
\references{
B{\"u}cher, A., & Kojadinovic, I. (2016).
A dependent multiplier bootstrap for the sequential empirical copula
process under strong mixing. \emph{Bernoulli}, \bold{22}(2), 927--968.
}
\seealso{
\code{\link{getMultipliers-MovingAverageMultipliers}}
}
\keyword{S4-classes}

