% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-MovingAverageMultipliers.R
\name{MovingAverageMultipliers-constructor}
\alias{MovingAverageMultipliers-constructor}
\alias{movingAverageMultipliers}
\title{Create an instance of the \code{\link{MovingAverageMultipliers}} class.}
\usage{
movingAverageMultipliers(l, N, kappa, innovDistr = rnorm)
}
\arguments{
\item{l}{the range of dependence of the multipliers (they are l-dependent);
for this scheme l needs to be chosen as an odd number and >= 3}

\item{N}{length of the sequence of dependent multipliers}

\item{kappa}{function from Section 5.2.1 (B{\"u}cher and Kojadinovic, 2016)}

\item{distrInnov}{function with one argument n that generates the independent
innovations}
}
\value{
Returns an instance of \code{MovingAverageMultipliers}.
}
\description{
Create an instance of the \code{\link{MovingAverageMultipliers}} class.
}
\keyword{Constructors}

