\name{MovingBlocks-constructor}
\alias{MovingBlocks-constructor}
\alias{movingBlocks}
\title{Create an instance of the \code{\link{MovingBlocks}} class.}
\usage{
movingBlocks(l, N)
}
\arguments{
  \item{l}{the block length for the block bootstrap
  methods}

  \item{N}{number of available observations to bootstrap
  from}
}
\value{
Returns an instance of \code{MovingBlocks}.
}
\description{
Create an instance of the \code{\link{MovingBlocks}} class.
}
\keyword{Constructors}

