% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-NoneMultipliers.R
\docType{class}
\encoding{latin1}
\name{NoneMultipliers-class}
\alias{NoneMultipliers}
\alias{NoneMultipliers-class}
\title{Class for Multipliers that are 1.}
\description{
\code{NoneMultipliers} is an S4 class that implements the not using multipliers
}
\details{
\code{NoneMultipliers} extends the S4 class
\code{\link{BootMultipliers}} and the remarks made in its documentation
apply here as well.

TODO: Add more description here!
}
\seealso{
\code{\link{getMultipliers-NoneMultipliers}}
}
\keyword{S4-classes}

