% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-NoneMultipliers.R
\name{NoneMultipliers-constructor}
\alias{NoneMultipliers-constructor}
\alias{noneMultipliers}
\title{Create an instance of the \code{NoneMultipliers} class.}
\usage{
noneMultipliers(N)
}
\arguments{
\item{N}{length of the sequence of dependent multipliers}
}
\value{
Returns an instance of \code{NoneMultipliers}.
}
\description{
l will be chosen as 1 by default.
}
\keyword{Constructors}

