% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-NonoverlappingBlocks.R
\docType{class}
\name{NonoverlappingBlocks-class}
\alias{NonoverlappingBlocks}
\alias{NonoverlappingBlocks-class}
\title{Class for Nonoverlapping Blocks Bootstrap implementation.}
\description{
\code{NonoverlappingBlocks} is an S4 class that implements the
nonoverlapping blocks bootstrap described in Carlstein (1986).
}
\details{
\code{NonoverlappingBlocks} extends the S4 class
\code{\link{BootPos}} and the remarks made in its documentation
apply here as well.

The Nonoverlapping Blocks Bootstrap method of Carlstein (1986) resamples blocks
randomly, with replacement from the collection of nonoverlapping blocks of
length \code{l} that start with observation \eqn{(i-1)l + 1}, where 
\eqn{i = 1, \ldots \lfloor N / l \rfloor}.
A more precise description of the procedure can also be found in
Lahiri (1999), p. 389.
}
\references{
Carlstein, E. (1986). The use of subseries methods for estimating the
variance of a general statistic from a stationary time series.
\emph{The Annals of Statistics}, \bold{14}, 1171--1179.
}
\seealso{
\code{\link{getPositions-NonoverlappingBlocks}}
}
\keyword{S4-classes}

