% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-NonoverlappingBlocks.R
\name{NonoverlappingBlocks-constructor}
\alias{NonoverlappingBlocks-constructor}
\alias{nonoverlappingBlocks}
\title{Create an instance of the \code{\link{NonoverlappingBlocks}} class.}
\usage{
nonoverlappingBlocks(l, N)
}
\arguments{
\item{l}{the block length for the block bootstrap methods}

\item{N}{number of available observations to bootstrap from}
}
\value{
Returns an instance of \code{NonoverlappingBlocks}.
}
\description{
Create an instance of the \code{\link{NonoverlappingBlocks}} class.
}
\keyword{Constructors}

