\name{quantspec-deprecated}
\alias{quantspec-deprecated}
\title{Deprecated functions in package quantspec}
\usage{
ct(i1, i2, n)

LaplacePeriodogram(X, taus, omegas = 1:(ceiling(length(X)/2) - 1),
  fromRanks = TRUE, showProgressBar = FALSE)

plotLaplacePeriodogram(LPG, taus, F = 1:length(LPG[, 1]),
  CL = 1:length(taus), hRange = FALSE, hOffset = FALSE,
  ylabel = expression({     {         hat(f)     }[n]^{         list(tau[1],
  tau[2])     } }(omega)), oma = c(2.5, 2.5, 2.5, 2.5), mar = c(4.5, 4.5, 1,
  0) + 0.1, cex.lab = 1.5)

smoothedLaplacePeriodogram(LPG, taus, W)
}
\arguments{
  \item{i1}{Parameter of DEPRECATED function.}

  \item{i2}{Parameter of DEPRECATED function.}

  \item{n}{Parameter of DEPRECATED function.}

  \item{X}{Parameter of DEPRECATED function.}

  \item{taus}{Parameter of DEPRECATED function.}

  \item{omegas}{Parameter of DEPRECATED function.}

  \item{fromRanks}{Parameter of DEPRECATED function.}

  \item{showProgressBar}{Parameter of DEPRECATED function.}

  \item{LPG}{Parameter of DEPRECATED function.}

  \item{F}{Parameter of DEPRECATED function.}

  \item{CL}{Parameter of DEPRECATED function.}

  \item{hRange}{Parameter of DEPRECATED function.}

  \item{hOffset}{Parameter of DEPRECATED function.}

  \item{ylabel}{Parameter of DEPRECATED function.}

  \item{oma}{Parameter of DEPRECATED function.}

  \item{mar}{Parameter of DEPRECATED function.}

  \item{cex.lab}{Parameter of DEPRECATED function.}

  \item{W}{Parameter of DEPRECATED function.}
}
\description{
These functions have been deprecated since Version 1.0.
They will be defunct in the near future!
}
\keyword{Deprecated}

