% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-SpecDistrWeight.R
\name{SpecDistrWeight-constructor}
\alias{SpecDistrWeight-constructor}
\alias{specDistrWeight}
\title{Create an instance of the \code{\link{SpecDistrWeight}} class.}
\usage{
specDistrWeight(descr = "Spectral Distribution Weights")
}
\arguments{
\item{descr}{a description for the weight object}
}
\value{
an instance of \code{SpecDistrWeight}.
}
\description{
Create an instance of the \code{\link{SpecDistrWeight}} class.
}
\examples{
wgt <- specDistrWeight()
}
\keyword{Constructors}

