% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-StationaryBlocks.R
\docType{class}
\name{StationaryBlocks-class}
\alias{StationaryBlocks}
\alias{StationaryBlocks-class}
\title{Class for Stationary Bootstrap implementation.}
\description{
\code{StationaryBlocks} is an S4 class that implements the moving blocks
bootstrap described in Politis and Romano (1994).
}
\details{
\code{StationaryBlocks} extends the S4 class
\code{\link{BootPos}} and the remarks made in its documentation
apply here as well.

The Stationary Bootstrap method of Politis and Romano (1994) resamples blocks
randomly, with replacement from the collection of overlapping blocks of
random length \eqn{L}, where \eqn{L} is chosen to have a geometric distribution
with parameter \code{1/l}. The first observation of every block is chosen
randomly from 1, 2, \ldots, \code{N}.
In contrast to \code{\link{MovingBlocks}} and \code{\link{NonoverlappingBlocks}},
the Circular Blocks Bootstrap uses elements from the periodically extended series.
A more precise description of the procedure can also be found in
Lahiri (1999), p. 390.
}
\references{
Politis, D. and Romano, J. P. (1994). The stationary bootstrap.
\emph{Journal of the American Statistical Association}, \bold{89}, 1303--1313.
}
\seealso{
\code{\link{getPositions-StationaryBlocks}}
}
\keyword{S4-classes}

