% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-StationaryBlocks.R
\name{StationaryBlocks-constructor}
\alias{StationaryBlocks-constructor}
\alias{stationaryBlocks}
\title{Create an instance of the \code{\link{StationaryBlocks}} class.}
\usage{
stationaryBlocks(l, N)
}
\arguments{
\item{l}{the block length for the block bootstrap methods}

\item{N}{number of available observations to bootstrap from}
}
\value{
Returns an instance of \code{StationaryBlocks}.
}
\description{
Create an instance of the \code{\link{StationaryBlocks}} class.
}
\keyword{Constructors}

