% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ts-models-mv}
\alias{mvts1}
\alias{mvts2}
\alias{mvts3}
\alias{mvts4}
\alias{ts-models-mv}
\title{Functions to simulate multivariate time series models from B{\"u}cher and
Kojadinovic (2016)}
\usage{
mvts1(n = 100, innov = c("Clayton1", "Clayton4", "Gumbel15", "Gumbel3"))

mvts2(n = 100, innov = c("Clayton1", "Clayton4", "Gumbel15", "Gumbel3"))

mvts3(n = 100, innov = c("Clayton1", "Clayton4", "Gumbel15", "Gumbel3"))

mvts4(n = 100, innov = c("Clayton1", "Clayton4", "Gumbel15", "Gumbel3"))
}
\arguments{
\item{n}{length of the time series to be returned}

\item{innov}{flag to indicate whether the independent innovations are
distributed according to Clayton copula with parameters 1 or 4
or according to Gumbel copula with parameters 1.5 or 3.}
}
\description{
Functions to simulate multivariate time series models from B{\"u}cher and
Kojadinovic (2016)
}
\details{
In these bivariate time series models the cross sectional dependence is
introduced via a cross sectional dependence of the innovations. More precisely,
\eqn{(\varepsilon_{i1}, \varepsilon_{i2}) := (\Phi^{-1}(U_{i1}), \Phi^{-1}(U_{i2}))},
with \eqn{\Phi} the c.d.f. of the standard normal distribution and
\eqn{(U_{i1}, U_{i2})} bivariate i.i.d. from either a Clayton copula with
parameters \eqn{1} or \eqn{4} or Gumbel copula \eqn{1.5} or \eqn{3},
respectively.

The four models then have the following mechanisms to introduce serial
dependence: 

\code{mvts1} bivariate AR(1) model:
\deqn{X_{ij} = 0.5 X_{i-1,j} + \varepsilon_{ij},}

\code{mvts2} bivariate NAR(1) model:
\deqn{X_{ij} = 0.6 sin(X_{i-1,j}) + \varepsilon_{ij},}

\code{mvts3} bivariate EXPAR(1) model:
\deqn{X_{ij} = (0.8 - 1.1 exp(-50 X^2_{i-1,j}) X_{i-1,j} + 0.1 \varepsilon_{ij},}

\code{mvts4} bivariate GARCH(1,1) model:
\deqn{X_{ij} = \sigma_{ij} \varepsilon_{ij}}
\deqn{\varepsilon_{i1}^2 = 0.012 + 0.919 \varepsilon_{i-1,1}^2 + 0.072 \varepsilon_{i-1,1}^2,}
#' and
\deqn{\varepsilon_{i2}^2 = 0.037 + 0.868 \varepsilon_{i-1,2}^2 + 0.115 \varepsilon_{i-1,2}^2.}
}
\examples{
require("zoo")
Y <- zoo(mvts1(100, innov = "Clayton4"))
plot(Y)
require("zoo")
Y <- zoo(mvts2(100, innov = "Gumbel15"))
plot(Y)
require("zoo")
Y <- zoo(mvts3(100, innov = "Gumbel15"))
plot(Y)
require("zoo")
Y <- zoo(mvts4(100, innov = "Gumbel15"))
plot(Y)
}
\references{
B{\"u}cher, A., & Kojadinovic, I. (2016).
A dependent multiplier bootstrap for the sequential empirical copula
process under strong mixing. \emph{Bernoulli}, \bold{22}(2), 927--968.
}

