% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-Weight.R
\docType{class}
\name{Weight-class}
\alias{Weight}
\alias{Weight-class}
\title{Interface Class to access different types of weighting functions.}
\description{
\code{Weights} is an S4 class that provides a common interface to
implementations of a weighting function \eqn{W_n(\omega)}{Wn(w)}.
}
\details{
Currently three implementations are available:
    (1) \code{\link{KernelWeight}},
    (2) \code{\link{LagKernelWeight}} and
    (3) \code{\link{SpecDistrWeight}}.
}
\section{Slots}{

\describe{
\item{\code{values}}{an array containing the weights.}

\item{\code{descr}}{a description to be used in some plots.}
}}
\keyword{S4-classes}

