% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-ClippedCov.R
\name{ClippedCov-constructor}
\alias{ClippedCov-constructor}
\alias{clippedCov}
\title{Create an instance of the \code{\link{ClippedCov}} class.}
\usage{
clippedCov(Y, maxLag = length(Y) - 1, levels.1 = c(0.5),
  levels.2 = levels.1, isRankBased = TRUE)
}
\arguments{
\item{Y}{Time series to calculate the copula covariance from}

\item{maxLag}{maximum lag between observations that should be used}

\item{levels.1}{a vector of numerics that determines the level of clipping}

\item{levels.2}{a vector of numerics that determines the level of clipping}

\item{isRankBased}{If true the time series is first transformed to pseudo data.}
}
\value{
Returns an instance of \code{ClippedCov}.
}
\description{
Create an instance of the \code{\link{ClippedCov}} class.
}
\examples{
ccf <- clippedCov(rnorm(100), maxLag = 10, levels.1 =c(0.1,0.5,0.9))
dim(getValues(ccf))
#print values for levels (.5,.5)
print(getValues(ccf)[,2,2])
}
\seealso{
\code{\link{LagOperator}}
}
\keyword{Constructors}

