% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-CovarianceMatrixMultipliers.R
\name{CovarianceMatrixMultipliers-constructor}
\alias{CovarianceMatrixMultipliers-constructor}
\alias{covarianceMatrixMultipliers}
\title{Create an instance of the \code{\link{CovarianceMatrixMultipliers}} class.}
\usage{
covarianceMatrixMultipliers(l, N, phi, distrInnov)
}
\arguments{
\item{l}{the range of dependence of the multipliers (they are l-dependent)}

\item{N}{length of the sequence of dependent multipliers}

\item{phi}{function from assumption (M3)}

\item{distrInnov}{function with one argument n that generates the independent
innovations}
}
\value{
Returns an instance of \code{CovarianceMatrixMultipliers}.
}
\description{
Create an instance of the \code{\link{CovarianceMatrixMultipliers}} class.
}
\keyword{Constructors}

