% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data-sp500}
\alias{data-sp500}
\alias{sp500}
\title{S&P 500: Standard and Poor's 500 stock index, 2007--2010}
\format{A univariate time series with 1008 observations; a \code{zoo} object}
\description{
Contains the returns of the S&P 500 stock index for the years 2007--2010.
The returns were computed as \code{(Adjusted.Close-Open)/Open}.
}
\details{
The data was downloaded from the Yahoo! Finance Website.
}
\examples{
plot(sp500)
}
\references{
Yahoo! Finance Website
\url{http://finance.yahoo.com/q/hp?s=^GSPC+Historical+Prices}
}
\keyword{data}

