\docType{methods}
\name{getBootPos-FreqRep}
\alias{getBootPos,FreqRep-method}
\alias{getBootPos-FreqRep}
\title{Get associated \code{\link{BootPos}} from a
\code{\link{FreqRep}}.}
\usage{
\S4method{getBootPos}{FreqRep}(object)
}
\arguments{
  \item{object}{\code{FreqRep} from which to get the
  \code{\link{BootPos}}.}
}
\value{
Returns the \code{\link{BootPos}} object associated.
}
\description{
Get associated \code{\link{BootPos}} from a
\code{\link{FreqRep}}.
}
\keyword{Access-association-functions}

