\docType{methods}
\name{getBw-KernelWeight}
\alias{getBw,KernelWeight-method}
\alias{getBw-KernelWeight}
\title{Get attribute \code{bw} (bandwidth / scaling parameter used for smoothing)
from a \code{KernelWeight}.}
\usage{
\S4method{getBw}{KernelWeight}(object)
}
\arguments{
  \item{object}{\code{KernelWeight} from which to get the
  bandwidth \code{bw}.}
}
\value{
Returns the \code{bw} attribute.
}
\description{
Get attribute \code{bw} (bandwidth / scaling parameter used
for smoothing) from a \code{KernelWeight}.
}
\keyword{Access-functions}

