% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-LagKernelWeight.R
\docType{methods}
\name{getBw-LagKernelWeight}
\alias{getBw,LagKernelWeight-method}
\alias{getBw-LagKernelWeight}
\title{Get attribute \code{bw} (bandwidth / scaling parameter used for smoothing)
from a \code{LagKernelWeight}.}
\usage{
\S4method{getBw}{LagKernelWeight}(object)
}
\arguments{
\item{object}{\code{LagKernelWeight} from which to get the bandwidth \code{bw}.}
}
\value{
Returns the \code{bw} attribute.
}
\description{
Get attribute \code{bw} (bandwidth / scaling parameter used for smoothing)
from a \code{LagKernelWeight}.
}
\keyword{Access-functions}

