% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Class-Weight.R
\docType{methods}
\name{getDescr-Weight}
\alias{getDescr,Weight-method}
\alias{getDescr-Weight}
\title{Get attribute \code{descr} from a \code{Weight}.}
\usage{
\S4method{getDescr}{Weight}(object)
}
\arguments{
\item{object}{\code{Weight} from which to get the \code{descr}.}
}
\value{
Returns the \code{descr} attribute.
}
\description{
Get attribute \code{descr} from a \code{Weight}.
}
\keyword{Access-functions}

