\docType{methods}
\name{getIsRankBased-FreqRep}
\alias{getIsRankBased,FreqRep-method}
\alias{getIsRankBased-FreqRep}
\title{Get \code{isRankBased} from a \code{\link{FreqRep}} object}
\usage{
\S4method{getIsRankBased}{FreqRep}(object)
}
\arguments{
  \item{object}{\code{FreqRep} of which to get the
  \code{isRankBased}}
}
\value{
Returns the attribute \code{isRankBased} that's a slot of
\code{object}.
}
\description{
Get \code{isRankBased} from a \code{\link{FreqRep}} object
}
\keyword{Access-functions}

