% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-LagOperator.R
\docType{methods}
\name{getMaxLag-LagOperator}
\alias{getMaxLag,LagOperator-method}
\alias{getMaxLag-LagOperator}
\title{Get \code{maxLag} from a \code{\link{LagOperator}} object.}
\usage{
\S4method{getMaxLag}{LagOperator}(object)
}
\arguments{
\item{object}{\code{LagOperator} of which to get the \code{maxLag}}
}
\value{
Returns the attribute \code{maxLag} that's a slot of \code{object}.
}
\description{
Get \code{maxLag} from a \code{\link{LagOperator}} object.
}
\keyword{Access-functions}

