% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-FreqRep.R
\docType{methods}
\name{getY-FreqRep}
\alias{getY,FreqRep-method}
\alias{getY-FreqRep}
\title{Get \code{Y} from a \code{\link{FreqRep}} object.}
\usage{
\S4method{getY}{FreqRep}(object, d = 1)
}
\arguments{
\item{object}{\code{FreqRep} of which to get the \code{Y}}

\item{d}{optional parameter that determine which time series to return;
                    may be a vector of elements 1, ..., D}
}
\value{
Returns the attribute \code{Y} that's a slot of \code{object}.
}
\description{
Get \code{Y} from a \code{\link{FreqRep}} object.
}

