% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-LagOperator.R
\docType{methods}
\name{plot-LagOperator}
\alias{plot,LagOperator,ANY-method}
\alias{plot-LagOperator}
\title{Plot the values of the \code{\link{LagOperator}}.}
\usage{
\S4method{plot}{LagOperator,ANY}(x)
}
\arguments{
\item{x}{The \code{\link{LagOperator}} to plot.}
}
\value{
Plots the \code{\link{LagOperator}}.
}
\description{
Not yet implemented.
}

